Move tpto command to shared.
