+++ Robust.Server/Console/Commands/PlayerCommands.cs
@@ -1,102 +1,13 @@
 using System;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text;
 using Robust.Server.Player;
 using Robust.Shared.Console;
-using Robust.Shared.Enums;
-using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
 using Robust.Shared.Network;
 
 namespace Robust.Server.Console.Commands
 {
-    public sealed class TeleportToCommand : LocalizedCommands
-    {
-        [Dependency] private readonly IPlayerManager _players = default!;
-        [Dependency] private readonly IEntityManager _entities = default!;
-
-        public override string Command => "tpto";
-
-        public override void Execute(IConsoleShell shell, string argStr, string[] args)
-        {
-            if (args.Length == 0)
-                return;
-
-            var target = args[^1];
-
-            if (!TryGetTransformFromUidOrUsername(target, shell, _entities, _players, out _, out var targetTransform))
-                return;
-
-            var transformSystem = _entities.System<SharedTransformSystem>();
-            var targetCoords = targetTransform.Coordinates;
-
-            if (args.Length == 1)
-            {
-                var player = shell.Player as IPlayerSession;
-                if (player?.Status != SessionStatus.InGame)
-                {
-                    shell.WriteError("You need to be in game to teleport to an entity.");
-                    return;
-                }
-
-                if (!_entities.TryGetComponent(player.AttachedEntity, out TransformComponent? playerTransform))
-                {
-                    shell.WriteError("You don't have an entity.");
-                    return;
-                }
-
-                transformSystem.SetCoordinates(player.AttachedEntity!.Value, targetCoords);
-                playerTransform.AttachToGridOrMap();
-            }
-            else
-            {
-                foreach (var victim in args)
-                {
-                    if (victim == target)
-                        continue;
-
-                    if (!TryGetTransformFromUidOrUsername(victim, shell, _entities, _players, out var uid, out var victimTransform))
-                        return;
-
-                    transformSystem.SetCoordinates(uid.Value, targetCoords);
-                    victimTransform.AttachToGridOrMap();
-                }
-            }
-        }
-
-        private static bool TryGetTransformFromUidOrUsername(
-            string str,
-            IConsoleShell shell,
-            IEntityManager entMan,
-            IPlayerManager playerMan,
-            [NotNullWhen(true)] out EntityUid? victimUid,
-            [NotNullWhen(true)] out TransformComponent? transform)
-        {
-            if (EntityUid.TryParse(str, out var uid) && entMan.TryGetComponent(uid, out transform))
-            {
-                victimUid = uid;
-                return true;
-            }
-
-            if (playerMan.TryGetSessionByUsername(str, out var session)
-                && entMan.TryGetComponent(session.AttachedEntity, out transform))
-            {
-                victimUid = session.AttachedEntity;
-                return true;
-            }
-
-            if (session == null)
-                shell.WriteError("Can't find username/id: " + str);
-            else
-                shell.WriteError(str + " does not have an entity.");
-
-            transform = null;
-            victimUid = default;
-            return false;
-        }
-    }
-
     public sealed class ListPlayers : LocalizedCommands
     {
         [Dependency] private readonly IPlayerManager _players = default!;
