+++ Robust.Shared/Console/Commands/TeleportCommands.cs
@@ -1,9 +1,15 @@
+using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
+using System.Linq;
 using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
+using Robust.Shared.Localization;
 using Robust.Shared.Map;
 using Robust.Shared.Map.Components;
 using Robust.Shared.Maths;
+using Robust.Shared.Players;
+using Robust.Shared.Utility;
 
 namespace Robust.Shared.Console.Commands;
 
@@ -59,6 +65,106 @@ internal sealed class TeleportCommand : LocalizedCommands
         }
 
         shell.WriteLine($"Teleported {shell.Player} to {mapId}:{posX},{posY}.");
+    }
+}
+
+public sealed class TeleportToCommand : LocalizedCommands
+{
+    [Dependency] private readonly ISharedPlayerManager _players = default!;
+    [Dependency] private readonly IEntityManager _entities = default!;
+
+    public override string Command => "tpto";
+    public override bool RequireServerOrSingleplayer => true;
+
+    public override void Execute(IConsoleShell shell, string argStr, string[] args)
+    {
+        if (args.Length == 0)
+            return;
+
+        var target = args[^1];
+
+        if (!TryGetTransformFromUidOrUsername(target, shell, _entities, _players, out _, out var targetTransform))
+            return;
+
+        var transformSystem = _entities.System<SharedTransformSystem>();
+        var targetCoords = targetTransform.Coordinates;
+
+        if (args.Length == 1)
+        {
+            var ent = shell.Player?.AttachedEntity;
+
+            if (!_entities.TryGetComponent(ent, out TransformComponent? playerTransform))
+            {
+                shell.WriteError(Loc.GetString("cmd-failure-no-attached-entity"));
+                return;
+            }
+
+            transformSystem.SetCoordinates(ent.Value, targetCoords);
+            playerTransform.AttachToGridOrMap();
+        }
+        else
+        {
+            foreach (var victim in args)
+            {
+                if (victim == target)
+                    continue;
+
+                if (!TryGetTransformFromUidOrUsername(victim, shell, _entities, _players, out var uid, out var victimTransform))
+                    return;
+
+                transformSystem.SetCoordinates(uid.Value, targetCoords);
+                victimTransform.AttachToGridOrMap();
+            }
+        }
+    }
+
+    private static bool TryGetTransformFromUidOrUsername(
+        string str,
+        IConsoleShell shell,
+        IEntityManager entMan,
+        ISharedPlayerManager playerMan,
+        [NotNullWhen(true)] out EntityUid? victimUid,
+        [NotNullWhen(true)] out TransformComponent? transform)
+    {
+        if (EntityUid.TryParse(str, out var uid) && entMan.TryGetComponent(uid, out transform))
+        {
+            victimUid = uid;
+            return true;
+        }
+
+        if (playerMan.Sessions.TryFirstOrDefault(x => x.ConnectedClient.UserName == str, out var session)
+            && entMan.TryGetComponent(session.AttachedEntity, out transform))
+        {
+            victimUid = session.AttachedEntity;
+            return true;
+        }
+
+        shell.WriteError(Loc.GetString("cmd-tpto-parse-error", ("str",str)));
+
+        transform = null;
+        victimUid = default;
+        return false;
+    }
+
+    public override CompletionResult GetCompletion(IConsoleShell shell, string[] args)
+    {
+        if (args.Length == 0)
+            return CompletionResult.Empty;
+        ;
+        var last = args[^1];
+
+        var users = _players.Sessions
+            .Select(x => x.ConnectedClient.UserName ?? string.Empty)
+            .Where(x => !string.IsNullOrWhiteSpace(x) && x.StartsWith(last, StringComparison.CurrentCultureIgnoreCase));
+
+        var hint = args.Length == 1 ? "cmd-tpto-destination-hint" : "cmd-tpto-victim-hint";
+        hint = Loc.GetString(hint);
+
+        var opts = CompletionResult.FromHintOptions(users, hint);
+        if (last != string.Empty && !EntityUid.TryParse(last, out _))
+            return opts;
+
+        return CompletionResult.FromHintOptions(opts.Options.Concat(CompletionHelper.EntityUids(last, _entities)), hint);
     }
 }
